新增DocIT, 添加doc api的相关测试”
